Initialisation EIR should now be set correctly for VectorTransmission (not tested).
git-svn-id: https://openmalaria.googlecode.com/svn/trunk@166 aa1ca794-fd11-11dd-ac65-f74f249cfaac

+++ include/global.h
@@ -79,7 +79,7 @@ public:
   /// temporal resolution of simulation, in days
   static int interval;
    //Simulation time steps per year
-  static int intervalsPerYear;
+  static size_t intervalsPerYear;
   // Maximum age of individuals in a scenario in time intervals
   static int maxAgeIntervals;
   static int simulationMode;
